src/requirements.txt
python-binance==1.0.17
flask==2.3.3
python-dotenv==1.0.1